feat: Added Bug DELETE Authentication

+++ BugReporterBe/src/bugs/providers/deleteBug.provider.ts
@@ -3,17 +3,20 @@ import { StatusCodes, ReasonPhrases } from "http-status-codes";
 import Bug from "../bugs.schema.ts";
 import errorLogger from "../../helpers/errorLogger.helper.ts";
 import { matchedData } from "express-validator";
+import Project from "../../projects/projects.schema.ts";
 
 async function deleteBugProvider(req: Request, res: Response)
 {
   const validatedResult = matchedData(req);
 
   try {
-
     const bugId = validatedResult.bugId;
+    const bug = await Bug.findOne({ _id: bugId });
     
+    const project = await Project.findOne({ _id: bug?.project, user: req.user?.sub })
+    if (!project) return res.status(StatusCodes.NOT_FOUND).json({ reason: "No Bug found for the provided ID" });
+
     const deletedBug = await Bug.deleteOne({ _id: bugId });
-    
     return res.status(StatusCodes.OK).json(deletedBug);
   }
   catch (error) {
